fix(money): BigInteger override FK columns, test ordering/uniqueness
Review fix round 1: payment_overrides.doc_a/doc_b now match documents.id's
BigInteger type (was Integer, silently truncatable). Adds tests proving the
doc_a < doc_b ordering check and the (kind, doc_a, doc_b) uniqueness
constraint are load-bearing, including a same-pair/different-kind case that
would catch a unique index accidentally scoped to just the pair.

+++ migrations/versions/0033_money_facts.py
@@ -52,10 +52,16 @@ def upgrade() -> None:
         sa.Column("id", sa.Integer(), primary_key=True),
         sa.Column("kind", sa.String(8), nullable=False),
         sa.Column(
-            "doc_a", sa.Integer(), sa.ForeignKey("documents.id", ondelete="CASCADE"), nullable=False
+            "doc_a",
+            sa.BigInteger(),
+            sa.ForeignKey("documents.id", ondelete="CASCADE"),
+            nullable=False,
         ),
         sa.Column(
-            "doc_b", sa.Integer(), sa.ForeignKey("documents.id", ondelete="CASCADE"), nullable=False
+            "doc_b",
+            sa.BigInteger(),
+            sa.ForeignKey("documents.id", ondelete="CASCADE"),
+            nullable=False,
         ),
         sa.Column(
             "created_at", sa.DateTime(timezone=True), nullable=False, server_default=sa.func.now()
